Fix uncaught exception thrown by boto3 (fixes #16)
While destroying clusters it might happen, that the DNS records are
already removed from route53. This should'nt prevent the osia from
continuing in cluster deletion. In order to fix this, it is necessary
to catch the exception thrown by boto3. It poses a bit of danger in case
of cluster installation, but the message and the error is properly logged,
so user should be informed that something bad happended with the DNS records.