Fix reference to cloud_env in README and config.py
The readme and error message in config.py have references to a config
option `base_env` for specifying the environment when one is not
specified on the command line. The config.py code, however, looks for a
config option called `cloud_env` (if I'm interpreting the code
correctly). This change updates the error message and README to use the
proper config option.