Allow setting worker replica count to 0
Some OpenShift cluster topologies (single node openshift, schedulable
master nodes) require creating a cluster with zero worker nodes. We want to be able
to deploy clusters with this configuration to more optimally use our
cloud resource quotas. This change is necessary to allow OSIA to
provision a cluster with zero workers. Prior to this change, if you set the count
to zero with the CLI `--worker-replicas` flag, Jinja would interpret
zero as False and therefore use the default value of 3. This moves the
logic for setting the default value to the base cloud class. I moved the master
node replica count logic to the same place for consistency.